# apps/backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_auto_generate_template_api_templates_auto_generate_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "file",
          "name"
        ],
        "title": "Body_auto_generate_template_api_templates_auto_generate_post",
        "type": "object"
      },
      "Body_create_template_api_templates_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "file",
          "name"
        ],
        "title": "Body_create_template_api_templates_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportFormatTemplatesUpdate": {
        "description": "種別のテンプレート構成を一括更新。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportFormatTemplateItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "ReportFormatTemplatesUpdate",
        "type": "object"
      },
...
      "TemplateAutoGenerateResponse": {
        "description": "POST /api/templates/auto-generate のレスポンス。",
        "properties": {
          "report": {
            "$ref": "#/components/schemas/TemplateAutoGenerateReportItem"
          },
          "template": {
            "$ref": "#/components/schemas/TemplateAutoGenerateTemplateItem"
          }
        },
        "required": [
          "template",
          "report"
        ],
        "title": "TemplateAutoGenerateResponse",
        "type": "object"
      },
...
      "TemplateSyncFileInput": {
        "description": "個別ファイル同期（インポート）のリクエスト。file_path は assets 基準の相対パス。",
        "properties": {
          "filePath": {
            "title": "Filepath",
            "type": "string"
          }
        },
        "required": [
          "filePath"
        ],
        "title": "TemplateSyncFileInput",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "現場報告管理システム Citadel のバックエンド API",
    "title": "Citadel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/report-formats/{format_id}/templates": {
      "get": {
        "description": "指定したレポート種別に紐づくテンプレート部品を sort_order 順で返す。",
        "operationId": "get_format_templates_api_report_formats__format_id__templates_get",
        "parameters": [
          {
            "in": "path",
            "name": "format_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Format Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Format Templates",
        "tags": [
          "report-formats"
        ]
      },
      "put": {
        "description": "指定したレポート種別のテンプレート構成を一括更新。既存の紐づけは削除し、body で指定した並びで再登録。",
        "operationId": "update_format_templates_api_report_formats__format_id__templates_put",
        "parameters": [
          {
            "in": "path",
            "name": "format_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Format Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportFormatTemplatesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Format Templates",
        "tags": [
          "report-formats"
        ]
      }
    },
...
    "/api/templates": {
      "get": {
        "description": "ReportTemplate（テンプレート部品）の一覧を返す。fileExists は実ファイルの存在を示す。",
        "operationId": "get_templates_api_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Templates",
        "tags": [
          "templates"
        ]
      },
      "post": {
        "description": "ブックバインダー：.xlsx を検疫し、通過したもののみ assets/templates に登録する（テンプレート部品として）。",
        "operationId": "create_template_api_templates_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_template_api_templates_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Template",
        "tags": [
          "templates"
        ]
      }
    },
    "/api/templates/auto-generate": {
      "post": {
        "description": "テンプレート .xlsx をアップロードし、AI でレポート情報を抽出したうえで\nプレースホルダを自動配置し、テンプレート部品として登録する。",
        "operationId": "auto_generate_template_api_templates_auto_generate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_auto_generate_template_api_templates_auto_generate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateAutoGenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auto Generate Template",
        "tags": [
          "templates"
        ]
      }
    },
...
    "/api/templates/sync-file": {
      "post": {
        "description": "指定した file_path（assets 基準の相対パス）のファイルを検疫し、\n通過した場合のみ DB に ReportTemplate を 1 件作成する。",
        "operationId": "sync_file_template_api_templates_sync_file_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateSyncFileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync File Template",
        "tags": [
          "templates"
        ]
      }
    },
    "/api/templates/{template_id}": {
      "delete": {
        "description": "指定IDのテンプレート部品をDBとファイルシステムから削除する。紐づく ReportFormatTemplate も削除。",
        "operationId": "delete_template_api_templates__template_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Template",
        "tags": [
          "templates"
        ]
      },
      "put": {
        "description": "指定IDのテンプレート部品のメタデータを更新する（ファイルは変更しない）。",
        "operationId": "update_template_api_templates__template_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Template",
        "tags": [
          "templates"
        ]
      }
    },
...
    "/api/templates/{template_id}/revalidate": {
      "post": {
        "description": "外部編集モード用。物理ファイルの存在確認・検疫を行い、問題なければ更新日時を返す。\nファイルがロックされている場合（Excel で開いている等）は 400 を返す。\nファイルが見つからない場合は new_file_path で再試行できる。",
        "operationId": "revalidate_template_api_templates__template_id__revalidate_post",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateRevalidateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revalidate Template",
        "tags": [
          "templates"
        ]
      }
    },
...
  }
}